functions/requirements.txt
# vvot29-face-detection
requests==2.28.1
boto3==1.24.75

# vvot29-boot
ydb==2.10.0
python-telegram-bot==13.15
pydantic==1.9.1
